Update key version for backward compatibility
This commit modifies the handling of key versions across multiple files in the MPC package. The `KeyInfo` struct now uses an `int` for the version field instead of `uint16`, and the `DefaultVersion` constant is updated from 0 to 1. Additionally, the `GetKeyInfoVersion` method in the `Node` struct now returns 0 for invalid key types and errors, improving error handling consistency. The `consumeResharingEvent` method in the `eventConsumer` is also updated to reflect these changes, ensuring better clarity and robustness in version management.

+++ pkg/keyinfo/keyinfo.go
@@ -11,7 +11,7 @@ import (
 type KeyInfo struct {
 	ParticipantPeerIDs []string `json:"participant_peer_ids"`
 	Threshold          int      `json:"threshold"`
-	Version            uint16   `json:"version"`
+	Version            int      `json:"version"`
 }
 
 type store struct {
